Change of flag to long and central particles to have - the label of the halo

+++ exshalos/exshalos/box.c
@@ -1,7 +1,7 @@
 #include "box.h"
 
 /*Generate a halo catalogue in a box from a given linear power spectrum*/
-size_t Generate_Halos_Box_from_Pk(fft_real *K, fft_real *P, int Nk, fft_real R_max, fft_real k_smooth, HALOS **halos, fft_real **posh, fft_real **velh, size_t *flag, fft_real *delta, fft_real *S, fft_real *V, int fixed, fft_real phase){
+size_t Generate_Halos_Box_from_Pk(fft_real *K, fft_real *P, int Nk, fft_real R_max, fft_real k_smooth, HALOS **halos, fft_real **posh, fft_real **velh, long *flag, fft_real *delta, fft_real *S, fft_real *V, int fixed, fft_real phase){
     int i, j;
     size_t nh;
     fft_complex *deltak;
@@ -68,7 +68,7 @@ size_t Generate_Halos_Box_from_Pk(fft_real *K, fft_real *P, int Nk, fft_real R_m
 }
 
 /*Generate a halo catalogue in a box from a given density grid*/
-size_t Generate_Halos_Box_from_Grid(fft_real *K, fft_real *P, int Nk, fft_real k_smooth, HALOS **halos, fft_real **posh, fft_real **velh, size_t *flag, fft_real *delta, fft_real *S, fft_real *V, int IN_disp){
+size_t Generate_Halos_Box_from_Grid(fft_real *K, fft_real *P, int Nk, fft_real k_smooth, HALOS **halos, fft_real **posh, fft_real **velh, long *flag, fft_real *delta, fft_real *S, fft_real *V, int IN_disp){
     int i, j;
     size_t ind, nh;
     fft_complex *deltak;
@@ -90,7 +90,7 @@ size_t Generate_Halos_Box_from_Grid(fft_real *K, fft_real *P, int Nk, fft_real k
     /*Attribute the inputed displacements to the halos*/
     if(IN_disp == TRUE){
         for(ind=0;ind<box.ng;ind++)
-            if(flag[ind] < box.ng){
+            if(flag[ind] < (long) box.ng){
                 (*posh)[3*flag[ind]] += S[3*ind];
                 (*posh)[3*flag[ind] + 1] += S[3*ind + 1];
                 (*posh)[3*flag[ind] + 2] += S[3*ind + 2];
